Plugins: renames plugin_loader
