+++ public/app/features/plugins/pluginLoader.ts
