+++ public/app/features/plugins/pluginPreloader.ts
@@ -7,7 +7,7 @@ import type { AppPluginConfig } from '@grafana/runtime';
 import { contextSrv } from 'app/core/services/context_srv';
 import { getPluginSettings } from 'app/features/plugins/pluginSettings';
 
-import { importAppPlugin } from './plugin_loader';
+import { importAppPlugin } from './pluginLoader';
 
 export type PluginPreloadResult = {
   pluginId: string;
