+++ public/app/features/variables/editor/getVariableQueryEditor.tsx
@@ -3,7 +3,7 @@ import { useCallback } from 'react';
 import { DataQuery, DataSourceApi, DataSourceJsonData, QueryEditorProps, StandardVariableQuery } from '@grafana/data';
 import { getTemplateSrv } from '@grafana/runtime';
 
-import { importDataSourcePlugin } from '../../plugins/plugin_loader';
+import { importDataSourcePlugin } from '../../plugins/pluginLoader';
 import {
   hasCustomVariableSupport,
   hasDatasourceVariableSupport,
